Address Extra unit tests
